signed/unsigned mismatch
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@8970 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/mod/formats/mod_tone_stream/mod_tone_stream.c
@@ -85,7 +85,7 @@ static switch_status_t silence_stream_file_read(switch_file_handle_t *handle, vo
 		return SWITCH_STATUS_FALSE;
 	}
 
-	if (*len > sh->samples) {
+	if (*len > (size_t)sh->samples) {
 		*len = sh->samples;
 	}
